refactor(forms): centralize forms
