+++ packages/client/app/components/forms/index.tsx
@@ -1,1 +1,3 @@
 export * from "./input";
+export * from "./login/login";
+export * from "./register/register";
