+++ packages/client/app/components/forms/login/login.scss
@@ -1,0 +1,4 @@
+@use "../../../theme.scss";
+
+.form-container {
+}
