+++ packages/client/app/components/forms/register/register.scss
@@ -1,0 +1,4 @@
+@use "../../../theme.scss";
+
+.form-container {
+}
