+++ packages/client/app/pages/login/login.tsx
@@ -1,3 +1,10 @@
+import { LoginForm } from "../../components/forms";
+import "./login.scss";
+
 export const LoginPage = () => {
-  return <div>LoginPage</div>;
+  return (
+    <div>
+      <LoginForm />
+    </div>
+  );
 };
